renames SigningStatusPanel -> SigningActions
