+++ src/layout/SigningActions/OnBehalfOfChooser.tsx
@@ -6,7 +6,7 @@ import { RadioButton } from 'src/components/form/RadioButton';
 import { Lang } from 'src/features/language/Lang';
 import { useLanguage } from 'src/features/language/useLanguage';
 import type { SigneeState } from 'src/layout/SigneeList/api';
-import type { AuthorizedOrganizationDetails } from 'src/layout/SigningStatusPanel/api';
+import type { AuthorizedOrganizationDetails } from 'src/layout/SigningActions/api';
 
 interface OnBehalfOfChooserProps {
   currentUserSignee: SigneeState | undefined;
