+++ src/layout/SigningActions/PanelAwaitingOtherSignatures.tsx
@@ -2,12 +2,12 @@ import React, { useEffect, useState } from 'react';
 
 import { Button } from 'src/app-components/Button/Button';
 import { Lang } from 'src/features/language/Lang';
-import { SigningPanel } from 'src/layout/SigningStatusPanel/PanelSigning';
+import { SigningPanel } from 'src/layout/SigningActions/PanelSigning';
 import { useNodeItem } from 'src/utils/layout/useNodeItem';
 import type { LayoutNode } from 'src/utils/layout/LayoutNode';
 
 type AwaitingOtherSignaturesPanelProps = {
-  node: LayoutNode<'SigningStatusPanel'>;
+  node: LayoutNode<'SigningActions'>;
   hasSigned: boolean;
 };
